Add basic testing structure
In addition to the requirements to run a test suite,
this change starts the process of migrating to a structure
where we have a callable (`pkg-update-deps`) that is
actually a symlink to a normal python file (`pkg_update_deps.py`).
This is done so that we can normally import the python file
in our unit tests, as well as potentially share some
functionality between the various callables.